fixed phpUnit tests
